+++ examples/src/main/kotlin/io/lambdarpc/examples/interactive_ml/Defs.kt
@@ -1,4 +1,4 @@
-package io.lambdarpc.examples.ml
+package io.lambdarpc.examples.interactive_ml
 
 import io.lambdarpc.utils.Endpoint
