+++ examples/src/main/kotlin/io/lambdarpc/examples/interactive_ml/dataservice/Lib.kt
@@ -1,0 +1,5 @@
+package io.lambdarpc.examples.interactive_ml.dataservice
+
+import io.lambdarpc.examples.interactive_ml.mlservice.Data
+
+fun dataloader(): Data = Data(1)
