+++ examples/src/main/kotlin/io/lambdarpc/examples/interactive_ml/mlservice/Lib.kt
@@ -1,4 +1,4 @@
-package io.lambdarpc.examples.ml.mlservice
+package io.lambdarpc.examples.interactive_ml.mlservice
 
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
