+++ examples/src/main/kotlin/io/lambdarpc/examples/interactive_ml/mlservice/Service.kt
@@ -1,9 +1,9 @@
-package io.lambdarpc.examples.ml.mlservice
+package io.lambdarpc.examples.interactive_ml.mlservice
 
 import io.lambdarpc.dsl.LibService
-import io.lambdarpc.examples.ml.mlEndpoint
-import io.lambdarpc.examples.ml.mlservice.facade.fit
-import io.lambdarpc.examples.ml.mlservice.facade.mlServiceId
+import io.lambdarpc.examples.interactive_ml.mlEndpoint
+import io.lambdarpc.examples.interactive_ml.mlservice.facade.fit
+import io.lambdarpc.examples.interactive_ml.mlservice.facade.mlServiceId
 
 fun main() {
     // Such service that looks like a library is called libservice
