--- examples/src/main/kotlin/io/lambdarpc/examples/ml/dataservice/Lib.kt
@@ -1,5 +1,0 @@
-package io.lambdarpc.examples.ml.dataservice
-
-import io.lambdarpc.examples.ml.mlservice.Data
-
-fun dataloader(): Data = Data(1)
